loading game from db, playing the game and switching players

+++ app/src/main/java/cz/mendelu/pef/dostihyasazky/model/SavedGame.kt
@@ -13,7 +13,7 @@ data class SavedGame(
     var date: String,
 
     @ColumnInfo(name = "player_on_turn_id")
-    var playerOnTurnId: Int
+    var playerOnTurnId: Long
 ){
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id_saved_game")
